# Optimize user tracking