--- a/atc/api/accessor/handler.go
+++ b/atc/api/accessor/handler.go
@@ -6,10 +6,15 @@ import (
 
 	"code.cloudfoundry.org/lager"
 	"github.com/concourse/concourse/atc/auditor"
-	"github.com/concourse/concourse/atc/db"
 )
 
 //go:generate counterfeiter net/http.Handler
+
+//go:generate coonuterfeiter UserTracker
+
+type UserTracker interface {
+	CreateOrUpdateUser(username, connector, sub string) error
+}
 
 func NewHandler(
 	logger lager.Logger,
@@ -17,7 +22,7 @@ func NewHandler(
 	accessFactory AccessFactory,
 	action string,
 	aud auditor.Auditor,
-	userFactory db.UserFactory,
+	userTracker UserTracker,
 ) http.Handler {
 	return accessorHandler{
 		logger:        logger,
@@ -25,7 +30,7 @@ func NewHandler(
 		accessFactory: accessFactory,
 		action:        action,
 		auditor:       aud,
-		userFactory:   userFactory,
+		userTracker:   userTracker,
 	}
 }
 
@@ -35,11 +40,10 @@ type accessorHandler struct {
 	accessFactory AccessFactory
 	action        string
 	auditor       auditor.Auditor
-	userFactory   db.UserFactory
+	userTracker   UserTracker
 }
 
 func (h accessorHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-
 	acc, err := h.accessFactory.Create(r, h.action)
 	if err != nil {
 		h.logger.Error("failed-to-create-accessor", err)
@@ -50,13 +54,13 @@ func (h accessorHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	claims := acc.Claims()
 
 	if claims.Sub != "" {
-		_, err = h.userFactory.CreateOrUpdateUser(
+		err = h.userTracker.CreateOrUpdateUser(
 			claims.UserName,
 			claims.Connector,
 			claims.Sub,
 		)
 		if err != nil {
-			h.logger.Error("failed-to-update-user-activity", err)
+			h.logger.Error("failed-to-track-user", err)
 			w.WriteHeader(http.StatusInternalServerError)
 			return
 		}