--- a/atc/api/teamserver/teamserver_suite_test.go
+++ b/atc/api/teamserver/teamserver_suite_test.go
@@ -1,0 +1,13 @@
+package teamserver_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestTeamserver(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Teamserver Suite")
+}