--- a/atc/atccmd/command.go
+++ b/atc/atccmd/command.go
@@ -17,6 +17,7 @@ import (
 
 	"code.cloudfoundry.org/clock"
 	"code.cloudfoundry.org/lager"
+	"github.com/MasterOfBinary/gobatch/batch"
 	"github.com/concourse/concourse"
 	"github.com/concourse/concourse/atc"
 	"github.com/concourse/concourse/atc/api"
@@ -1747,6 +1748,12 @@ func (cmd *RunCommand) constructAPIHandler(
 		cmd.Auditor.EnableVolumeAuditLog,
 		logger,
 	)
+
+	batcher := accessor.NewBatcher(logger, dbUserFactory, &batch.ConfigValues{
+		MaxItems: 100,
+		MinTime:  15 * time.Second,
+	})
+
 	apiWrapper := wrappa.MultiWrappa{
 		wrappa.NewAPIMetricsWrappa(logger),
 		wrappa.NewAPIAuthWrappa(
@@ -1756,7 +1763,7 @@ func (cmd *RunCommand) constructAPIHandler(
 			checkWorkerTeamAccessHandlerFactory,
 		),
 		wrappa.NewConcourseVersionWrappa(concourse.Version),
-		wrappa.NewAccessorWrappa(logger, accessFactory, aud, dbUserFactory),
+		wrappa.NewAccessorWrappa(logger, accessFactory, aud, batcher),
 		wrappa.NewCompressionWrappa(logger),
 	}