--- a/atc/atccmd/command.go
+++ b/atc/atccmd/command.go
@@ -652,19 +652,11 @@ func (cmd *RunCommand) constructAPIMembers(
 	dbClock := db.NewClock()
 	dbWall := db.NewWall(dbConn, &dbClock)
 
-	customRoles, err := cmd.parseCustomRoles(logger)
-	if err != nil {
-		return nil, err
-	}
-
-	cacher := accessor.NewCacher(logger, dbConn.Bus(), teamFactory)
+	tokenVerifier := cmd.constructTokenVerifier(httpClient)
 
 	accessFactory := accessor.NewAccessFactory(
-		cmd.constructTokenVerifier(httpClient),
-		cacher,
 		cmd.SystemClaimKey,
 		cmd.SystemClaimValues,
-		customRoles,
 	)
 
 	middleware := token.NewMiddleware(cmd.Auth.AuthFlags.SecureCookies)
@@ -689,6 +681,8 @@ func (cmd *RunCommand) constructAPIMembers(
 		credsManagers,
 		accessFactory,
 		dbWall,
+		tokenVerifier,
+		dbConn.Bus(),
 	)
 	if err != nil {
 		return nil, err
@@ -1095,7 +1089,7 @@ func (cmd *RunCommand) validateCustomRoles() error {
 	return nil
 }
 
-func (cmd *RunCommand) parseCustomRoles(logger lager.Logger) (map[string]string, error) {
+func (cmd *RunCommand) parseCustomRoles() (map[string]string, error) {
 	mapping := map[string]string{}
 
 	path := cmd.ConfigRBAC.Path()
@@ -1696,7 +1690,7 @@ func (cmd *RunCommand) constructLoginHandler(
 	return skyserver.NewSkyHandler(skyServer), nil
 }
 
-func (cmd *RunCommand) constructTokenVerifier(httpClient *http.Client) accessor.Verifier {
+func (cmd *RunCommand) constructTokenVerifier(httpClient *http.Client) accessor.TokenVerifier {
 
 	publicKeyPath, _ := url.Parse("/sky/issuer/keys")
 	publicKeyURL := cmd.ExternalURL.URL.ResolveReference(publicKeyPath)
@@ -1729,6 +1723,8 @@ func (cmd *RunCommand) constructAPIHandler(
 	credsManagers creds.Managers,
 	accessFactory accessor.AccessFactory,
 	dbWall db.Wall,
+	tokenVerifier accessor.TokenVerifier,
+	notifications db.NotificationsBus,
 ) (http.Handler, error) {
 
 	checkPipelineAccessHandlerFactory := auth.NewCheckPipelineAccessHandlerFactory(teamFactory)
@@ -1749,10 +1745,27 @@ func (cmd *RunCommand) constructAPIHandler(
 		logger,
 	)
 
-	batcher := accessor.NewBatcher(logger, dbUserFactory, &batch.ConfigValues{
-		MaxItems: 100,
-		MinTime:  15 * time.Second,
-	})
+	batcher := accessor.NewBatcher(
+		logger,
+		dbUserFactory,
+		&batch.ConfigValues{
+			MaxItems: 100,
+			MinTime:  15 * time.Second,
+		},
+	)
+
+	cacher := accessor.NewCacher(
+		logger,
+		notifications,
+		teamFactory,
+		time.Minute,
+		time.Minute,
+	)
+
+	customRoles, err := cmd.parseCustomRoles()
+	if err != nil {
+		return nil, err
+	}
 
 	apiWrapper := wrappa.MultiWrappa{
 		wrappa.NewAPIMetricsWrappa(logger),
@@ -1763,7 +1776,15 @@ func (cmd *RunCommand) constructAPIHandler(
 			checkWorkerTeamAccessHandlerFactory,
 		),
 		wrappa.NewConcourseVersionWrappa(concourse.Version),
-		wrappa.NewAccessorWrappa(logger, accessFactory, aud, batcher),
+		wrappa.NewAccessorWrappa(
+			logger,
+			accessFactory,
+			tokenVerifier,
+			cacher,
+			batcher,
+			aud,
+			customRoles,
+		),
 		wrappa.NewCompressionWrappa(logger),
 	}