--- a/atc/cache.go
+++ b/atc/cache.go
@@ -1,0 +1,6 @@
+package atc
+
+const (
+	TeamCacheName    = "teams"
+	TeamCacheChannel = "team_cache"
+)