# Why do we need this PR

This PR adds optimizations during the authorization flow for each api request. Every time concourse handles an api request it needs to pull down the list of all teams, so that it can iterate over the auth configuration to determine if the current user has access to that teams information. On every request it also tracks the user who made the request and writes this information to the db so that we can determine the last time the user was active in concourse (this is important for billing).

So this PR introduces two optimizations:
- caching the `GetTeams` call for 1 minute (which gets a db notification to invalidate on `SetTeam`)
- batching `CreateOrUpdateUser` writes to the db every 15s (or a batch size of 100 items)


# Benchmark results:

With batch and cache. Running 30s.
```
Running 30s test @ http://localhost:8080/api/v1/pipelines
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.62ms    3.60ms  46.72ms   84.30%
    Req/Sec   677.27     63.23   838.00     71.17%
  40466 requests in 30.02s, 10.88MB read
Requests/sec:   1348.02
Transfer/sec:    371.23KB
```

Master branch
```
Running 10s test @ http://localhost:8080/api/v1/pipelines
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     6.60ms    4.06ms  36.40ms   88.42%
    Req/Sec   812.27     85.47     0.97k    64.50%
  16174 requests in 10.00s, 4.35MB read
Requests/sec:   1617.05
Transfer/sec:    445.32KB
```

Without batch and cache `use-dex-token` branch
```
Running 10s test @ http://localhost:8080/api/v1/pipelines
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.93ms    7.44ms  83.86ms   79.78%
    Req/Sec   301.43     30.99   363.00     74.50%
  6007 requests in 10.01s, 1.62MB read
Requests/sec:    600.24
Transfer/sec:    165.30KB
```