@cirocosta 2020-04-14
@xtremerui do you think we should also benchmark against 5.8 / 6.0 before the tokens pr just to have a baseline of how we're doing against what we had previously?