@xtremerui 2020-04-14
yeah thats the plan. We need to know what improvement is now. The above benchmarks could be obslete already.