@cirocosta 2020-04-16
Hey @pivotal-jwinters @xtremerui , 

could you give a brief description of how we're accomplishing what we want here? i see the use of the `batch` library here, but, could you elaborate on how it does the sync? e.g., is it writing in batch based on a given interval? buffer getting full / ratio of some sorts?

thanks!