@cirocosta 2020-04-17
Hey,

I was doing some testing here of this, and wanted to share the results i've got.

The setup:

1. static rate of requests (1500/s), hold for 15s
2. `web` (varying its versions) and `db` (same version: 12.2) running in docker,
w/ no limits set, in the same machine
  - versions tested: 5.8.0, 6.0.0, master
    (5e315c6ccfe2fdb087ecb40a7a5339b5dc31feb5), and this PR
    (8f0d9efc84488fd61801125aeefd7df580b1486a)

3. Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz (turbo boost 5GHz)

Naturally, all four scenarios worked just fine hitting `/api/v1/info/pipelines`
with all of the requests authenticated.

```
5.8.0

        Requests      [total, rate, throughput]         22500, 1500.07, 1499.98
        Duration      [total, attack, wait]             15s, 14.999s, 906.663µs
        Latencies     [min, mean, 50, 90, 95, 99, max]  436.009µs, 931.217µs, 932.175µs, 1.259ms, 1.416ms, 2.358ms, 11.186ms
        Bytes In      [total, mean]                     67500, 3.00
        Bytes Out     [total, mean]                     0, 0.00
        Success       [ratio]                           100.00%
        Status Codes  [code:count]                      200:22500
        Error Set:


6.0.0
        Requests      [total, rate, throughput]         22500, 1500.08, 1500.03
        Duration      [total, attack, wait]             15s, 14.999s, 490µs
        Latencies     [min, mean, 50, 90, 95, 99, max]  433.87µs, 1.049ms, 995.682µs, 1.227ms, 1.413ms, 2.607ms, 88.059ms
        Bytes In      [total, mean]                     67500, 3.00
        Bytes Out     [total, mean]                     0, 0.00
        Success       [ratio]                           100.00%
        Status Codes  [code:count]                      200:22500
        Error Set:


master
        Requests      [total, rate, throughput]         22500, 1500.09, 1482.51
        Duration      [total, attack, wait]             15.177s, 14.999s, 177.851ms
        Latencies     [min, mean, 50, 90, 95, 99, max]  4.923ms, 198.958ms, 115.691ms, 412.016ms, 697.929ms, 1.458s, 4.69s
        Bytes In      [total, mean]                     67500, 3.00
        Bytes Out     [total, mean]                     0, 0.00
        Success       [ratio]                           100.00%
        Status Codes  [code:count]                      200:22500
        Error Set:


this pull request
        Requests      [total, rate, throughput]         22500, 1500.07, 1499.94
        Duration      [total, attack, wait]             15.001s, 14.999s, 1.243ms
        Latencies     [min, mean, 50, 90, 95, 99, max]  550.363µs, 976.507µs, 926.015µs, 1.296ms, 1.426ms, 2.318ms, 16.551ms
        Bytes In      [total, mean]                     67500, 3.00
        Bytes Out     [total, mean]                     0, 0.00
        Success       [ratio]                           100.00%
        Status Codes  [code:count]                      200:22500
        Error Set:
```

Note how `master`'s latencies are all over the place, from 4.93 millisecs up to
4.69 seconds - even its median is already higher than the max of any of the
other ones.

with regards to resource consumption, we're back to being very close to what we
have before too (look at the last bump there - this PR - and the previous -
`master`, where both DB and WEB utilization was higher).

![Screenshot from 2020-04-17 14-21-48](https://user-images.githubusercontent.com/3574444/79601555-fc172580-80b6-11ea-84f0-a5debc0b9217.png)


good job!!