@pivotal-jwinters 2020-04-17
@cirocosta I'm actually surprised that you got that throughput on `master`.

How many pipelines did you have configured? We noticed that the performance boost from this PR is much less noticeable once you have a certain number of pipelines (assuming you're hitting the /api/v1/pipelines endpoint)