@cirocosta 2020-04-17
@pivotal-jwinters oh, I had 0 pipelines set - my rationale is that we wanted to just benchmark the effect of the accessor :thinking: was that wrong?