@pivotal-jwinters 2020-04-17
@cirocosta I think that's a fair assumption. I just saw a much lower throughput on master than what you observed, so I was trying to rationalize.