@pivotal-jamie-klassen 2020-04-23
👍 to release note - https://github.com/concourse/concourse/blob/master/release-notes/latest.md#4950 could probably be expanded to mention the DB impact.